+++ pluginshared/ct_itemdrawable/ct_circle.cpp
@@ -62,12 +62,12 @@ QString CT_Circle::staticGetType()
     return CT_AbstractShape::staticGetType() + "/CT_Circle";
 }
 
-float CT_Circle::getRadius() const
+double CT_Circle::getRadius() const
 {
     return ((const CT_CircleData&)getData()).getRadius();
 }
 
-float CT_Circle::getError() const
+double CT_Circle::getError() const
 {
     return ((const CT_CircleData&)getData()).getError();
 }
@@ -110,42 +110,6 @@ CT_Circle* CT_Circle::staticCreateZAxisAlignedCircleFromPointCloud(const CT_OutA
 
     return new CT_Circle((const CT_OutAbstractSingularItemModel *)model, result, data);
 }
-
-// Code non vérifié...
-//void CT_Circle::computeBoundingBox()
-//{
-//    const CT_CircleData &data = (const CT_CircleData&)getData();
-
-//    const QVector3D &center = data.getCenter();
-//    const QVector3D &direction = data.getDirection();
-//    float radius = data.getRadius();
-
-//    QVector3D u(0,0,0);
-
-//    if(fabs(direction.x()) >= fabs(direction.y()))
-//    {
-//        double factor = 1.0/sqrt(direction.x()*direction.x()+direction.z()*direction.z());
-//        u.setX(-direction.z()*factor);
-//        u.setZ(direction.x()*factor);
-//    }
-//    else
-//    {
-//        double factor = 1.0/sqrt(direction.y()*direction.y()+direction.z()*direction.z());
-//        u.setY(direction.z()*factor);
-//        u.setZ(-direction.y()*factor);
-//    }
-
-//    QVector3D v = QVector3D::crossProduct(direction, u);
-//    v.normalize();
-
-//    _minCoordinates.setX(center.x() - radius*u.x() - radius*v.x());
-//    _minCoordinates.setY(center.y() - radius*u.y() - radius*v.y());
-//    _minCoordinates.setZ(center.z() - radius*u.z() - radius*v.z());
-
-//    _maxCoordinates.setX(center.x() + radius*u.x() + radius*v.x());
-//    _maxCoordinates.setY(center.y() + radius*u.y() + radius*v.y());
-//    _maxCoordinates.setZ(center.z() + radius*u.z() + radius*v.z());
-//}
 
 #ifdef USE_BOOST_OLD
 BOOST_CLASS_EXPORT_IMPLEMENT(CT_CircleData)
