+++ pluginshared/ct_itemdrawable/ct_circle2d.h
@@ -68,7 +68,7 @@ public:
     virtual QString getType() const;
     static QString staticGetType();
 
-    float getRadius() const;
+    double getRadius() const;
 
     virtual CT_AbstractItemDrawable* copy(const CT_OutAbstractItemModel *model, const CT_AbstractResult *result, CT_ResultCopyModeList copyModeList);
